Fix: Intl.DateTimeFormat weekday option error in getEtDayOfWeek - Changed weekday from 'numeric' to 'long' to fix 500 error for ACQ agent KPI cards - Resolves 'Value numeric out of range for Intl.DateTimeFormat options property weekday' error

+++ server/src/services/metricsService.ts
@@ -51,7 +51,7 @@ function getEtDayOfWeek(d: Date): number {
   // Get day of week in ET timezone (0 = Sunday, 1 = Monday, ..., 6 = Saturday)
   const formatter = new Intl.DateTimeFormat('en-US', {
     timeZone: 'America/New_York',
-    weekday: 'numeric',
+    weekday: 'long',
   });
   const dayName = formatter.format(d);
   const dayMap: Record<string, number> = {
